requirements/base.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
psycopg==3.1.9
channels==4.0.0
channels-redis==4.1.0
Pillow==10.0.0
celery==5.3.4
redis==5.0.1
django-treebeard==4.7
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
boto3==1.28.84
django-storages==1.14.2
